make Point use doubles, more graphical refactoring

+++ src/com/lushprojects/circuitjs1/client/Polygon.java
@@ -25,21 +25,21 @@ class Polygon {
 	
 	private static final int MIN_LENGTH = 4;
 	public int npoints;
-	public int xpoints[];
-	public int ypoints[];
+	public double xpoints[];
+	public double ypoints[];
 	
 	
 	public Polygon(){
 //		poly = new ArrayList<Point>();
-        xpoints = new int[MIN_LENGTH];
-        ypoints = new int[MIN_LENGTH];
+        xpoints = new double[MIN_LENGTH];
+        ypoints = new double[MIN_LENGTH];
 	}
 	
 //	public void addPoint(int x, int y){
 //		poly.add(new Point(x,y));
 //	}
 	
-    public void addPoint(int x, int y) {
+    public void addPoint(double x, double y) {
         if (npoints >= xpoints.length || npoints >= ypoints.length) {
             int newLength = npoints * 2;
             // Make sure that newLength will be greater than MIN_LENGTH and
@@ -61,8 +61,8 @@ class Polygon {
 //        }
     }
     
-    private int[] expand(int[] in, int newlen) {
-    	int[] out=new int[newlen];
+    private double[] expand(double[] in, int newlen) {
+    	double[] out=new double[newlen];
     	for(int i=0; i<in.length; i++)
     		out[i]=in[i];
     	return out;
